Use a timer to defer configuration
There is no way to establish precedence or priority of autocommands in
Neovim, yet we want to ensure that the editorconfig options are applied
last, after any file type settings or settings from other plugins, to
ensure they have the highest priority. Using a timer with value 0 causes
the function to be deferred until after all other autocommands have run.