Don't use a timer to set editorconfig settings
On some slower systems there is a visible delay before the timer
executes and the settings are applied. Instead of using a timer, use the
VimEnter autocommand. The VimEnter autocommand happens after the Buf*
events that we're interested in, so we need to call the config function
in the VimEnter autocommand itself, and then create the Buf*
autocommands.