Define autocommand before first redraw
Define the BufRead autocommand used by editorconfig just before the
first screen redraw (using a decoration provider) rather than in a
VimEnter autocommand. This is necessary to ensure the editorconfig
autocommand is defined as late as possible (so that it fires last and
thus overrides any other plugins). The issue with VimEnter is that the
VimEnter autocommand must also apply EditorConfig settings for already
open buffers, which overrides any modelines present in those buffers.
Using a decoration provider allows the autocommands to be defined
*before* the first BufRead event fires, so the first opened buffer's
modeline settings are applied after the editorconfig autocommand fires.